Add android example
