+++ examples/AccordionExample/android/app/src/main/java/com/accordionexample/MainActivity.java
@@ -27,13 +27,14 @@ public class MainActivity extends ReactActivity {
         return BuildConfig.DEBUG;
     }
 
-   /**
-   * A list of packages used by the app. If the app uses additional views
-   * or modules besides the default ones, add more packages here.
-   */
+    /**
+     * A list of packages used by the app. If the app uses additional views
+     * or modules besides the default ones, add more packages here.
+     */
     @Override
     protected List<ReactPackage> getPackages() {
-      return Arrays.<ReactPackage>asList(
-        new MainReactPackage());
+        return Arrays.<ReactPackage>asList(
+            new MainReactPackage()
+        );
     }
 }
